+++ app/Http/Controllers/Rtrw/TamuController.php
@@ -23,7 +23,7 @@ class TamuController extends Controller
     public $guard2 = 'satpam';
 
     function generateKode($tabel, $kolom_acuan, $prefix, $str_format){
-        $query = DB::connection('sqlsrv2')->select("select right(max($kolom_acuan), ".strlen($str_format).")+1 as id from $tabel where $kolom_acuan like '$prefix%'");
+        $query = DB::connection($this->sql)->select("select right(max($kolom_acuan), ".strlen($str_format).")+1 as id from $tabel where $kolom_acuan like '$prefix%'");
         $query = json_decode(json_encode($query),true);
         $kode = $query[0]['id'];
         $id = $prefix.str_pad($kode, strlen($str_format), $str_format, STR_PAD_LEFT);
@@ -188,11 +188,16 @@ class TamuController extends Controller
             return response()->json($success, $this->successStatus);     
         } catch (\Throwable $e) {
             DB::connection($this->sql)->rollback();
-            if(Storage::disk('s3')->exists('rtrw/'.$foto)){
-                Storage::disk('s3')->delete('rtrw/'.$foto);
+            if(isset($foto)){
+                if(Storage::disk('s3')->exists('rtrw/'.$foto)){
+                    Storage::disk('s3')->delete('rtrw/'.$foto);
+                }
             }
-            if(Storage::disk('s3')->exists('rtrw/'.$output_file)){
-                Storage::disk('s3')->delete('rtrw/'.$output_file);
+            
+            if(isset($output_file)){
+                if(Storage::disk('s3')->exists('rtrw/'.$output_file)){
+                    Storage::disk('s3')->delete('rtrw/'.$output_file);
+                }
             }
             $success['status'] = false;
             $success['message'] = "Data Tamu Masuk gagal disimpan ".$e;
